Finish docs for ocp dir

+++ leap_c/examples/chain/controller.py
@@ -146,7 +146,7 @@ class ChainController(ParameterizedController):
 
     @property
     def param_space(self) -> gym.Space:
-        return self.param_manager.get_param_space(dtype=np.float64)
+        return self.param_manager.get_param_space(dtype=np.float32)
 
     def default_param(self, obs) -> np.ndarray:
         return self.param_manager.learnable_parameters_default.cat.full().flatten()  # type:ignore
